refactor: add native ad without styles tests
