+++ packages/ad/__tests__/android/ad.android.test.js
@@ -1,4 +1,4 @@
-import shared from "../ad.native";
+import shared from "../ad.shared";
 
 describe("android", () => {
   shared();
